Refactoring package from feargames to novaverse.
