+++ pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>it.feargames</groupId>
+    <groupId>it.novaverse</groupId>
     <artifactId>commonutilities</artifactId>
     <version>2.3.0-SNAPSHOT</version>
     <packaging>jar</packaging>
