--- src/main/java/it/feargames/commonutilities/annotation/RegisterCommands.java
@@ -1,11 +1,0 @@
-package it.feargames.commonutilities.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target({ElementType.TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface RegisterCommands {
-}
