+++ src/main/java/it/novaverse/commonutilities/annotation/RegisterCommands.java
@@ -1,0 +1,11 @@
+package it.novaverse.commonutilities.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target({ElementType.TYPE})
+@Retention(RetentionPolicy.RUNTIME)
+public @interface RegisterCommands {
+}
