+++ src/main/java/it/novaverse/commonutilities/module/implementation/gameplay/DisableFireworkDamage.java
@@ -1,0 +1,34 @@
+package it.novaverse.commonutilities.module.implementation.gameplay;
+
+import it.novaverse.commonutilities.annotation.ConfigValue;
+import it.novaverse.commonutilities.annotation.RegisterListeners;
+import it.novaverse.commonutilities.module.Module;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+import org.bukkit.entity.EntityType;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.Listener;
+import org.bukkit.event.entity.EntityDamageByEntityEvent;
+
+@NoArgsConstructor(access = AccessLevel.PACKAGE)
+@RegisterListeners
+public class DisableFireworkDamage implements Module, Listener {
+
+    @ConfigValue
+    private Boolean enabled = false;
+
+    @Override
+    public boolean isEnabled() {
+        return enabled;
+    }
+
+    @EventHandler(ignoreCancelled = true, priority = EventPriority.LOW)
+    public void onDamageByEntity(EntityDamageByEntityEvent event) {
+        if (event.getDamager().getType() != EntityType.FIREWORK) {
+            return;
+        }
+        event.setCancelled(true);
+    }
+
+}
