+++ src/main/java/it/novaverse/commonutilities/module/implementation/gameplay/DisableItemSwap.java
@@ -1,0 +1,30 @@
+package it.novaverse.commonutilities.module.implementation.gameplay;
+
+import it.novaverse.commonutilities.annotation.ConfigValue;
+import it.novaverse.commonutilities.annotation.RegisterListeners;
+import it.novaverse.commonutilities.module.Module;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.Listener;
+import org.bukkit.event.player.PlayerSwapHandItemsEvent;
+
+@NoArgsConstructor(access = AccessLevel.PACKAGE)
+@RegisterListeners
+public class DisableItemSwap implements Module, Listener {
+
+    @ConfigValue
+    private Boolean enabled = true;
+
+    @Override
+    public boolean isEnabled() {
+        return enabled;
+    }
+
+    @EventHandler(ignoreCancelled = true, priority = EventPriority.LOW)
+    public void onItemSwap(PlayerSwapHandItemsEvent event) {
+        event.setCancelled(true);
+    }
+
+}
