+++ src/main/java/it/novaverse/commonutilities/module/implementation/gameplay/combat/DisableAttackCooldown.java
@@ -1,0 +1,31 @@
+package it.novaverse.commonutilities.module.implementation.gameplay.combat;
+
+import it.novaverse.commonutilities.annotation.ConfigValue;
+import it.novaverse.commonutilities.annotation.RegisterListeners;
+import it.novaverse.commonutilities.module.Module;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+import org.bukkit.attribute.Attribute;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.Listener;
+import org.bukkit.event.player.PlayerJoinEvent;
+
+@NoArgsConstructor(access = AccessLevel.PACKAGE)
+@RegisterListeners
+public class DisableAttackCooldown implements Module, Listener {
+
+    @ConfigValue
+    private Boolean enabled = false;
+
+    @Override
+    public boolean isEnabled() {
+        return enabled;
+    }
+
+    @EventHandler(priority = EventPriority.HIGH)
+    public void onPlayerJoin(PlayerJoinEvent event) {
+        event.getPlayer().getAttribute(Attribute.GENERIC_ATTACK_SPEED).setBaseValue(32.0D);
+    }
+
+}
