+++ src/main/resources/plugin.yml
@@ -1,10 +1,10 @@
 name: CommonUtilities
 version: ${project.version}
-main: it.feargames.commonutilities.CommonUtilities
+main: it.novaverse.commonutilities.CommonUtilities
 authors: [ sgdc3 ]
 description: A collection of common commands and features
 softdepend: [ packetevents ]
-api-version: '1.17'
+api-version: '1.20'
 
 
 libraries:
